requirements.txt
dj-database-url==0.5.0
Django==2.2.7
django-filter==2.3.0
gunicorn==20.0.4
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
